Add namespaces to internal macro calls in case of external usage
In Boston, calling the source_edfi3 macro from the dbt project importing this
package results in dbt not resolving the call to the json_ci_get macro. Adding
the namespace fixes this. Pre-emptively also adding the namespace to calls to
json_flatten inside the flatten_arrays macro in case that gets used by
importing projects as well.